requirements.txt
psutil==5.9.5
netaddr==0.8.0
python-nmap==0.7.1